--- src/components/filters/index.ts
@@ -1,8 +1,0 @@
-export * from './components/filters-container';
-export * from './components/filters-section';
-export * from './components/filters-list';
-export * from './components/filters-checkbox';
-export * from './components/filters-tag';
-export * from './components/filters-tag-wrapper';
-export * from './helpers';
-export * from './hooks/useFacetsData';
