+++ src/views/search-results-page/components/filters/components/filters-checkbox.tsx
