+++ src/views/search-results-page/components/filters/components/filters-date-slider/helpers.ts
