+++ src/views/search-results-page/components/filters/components/filters-list.tsx
